Add initial implementation of the TwiML generation package, twiml
Twilio has their own XML-based Markup Language called TwiML. This markup is used
when sending and receiving phone calls using Twilio. It gives Twilio
instructions on how to handle and route the call, as well as what to do once the
call has been answered (such as saying something).

This package introduces the ability to render all aspects of TwiML to allow
users to do this in a way that feels idiomatic.